Code corrections for the Implementation Backend
These are minor code corrections for the implementation of the project backend, specifically for the clustering.
Many changes only affect whitepsaces - these are for eclipse compliance and can be ignored.
The membership algorithm has changed. The previous calculation would have let to an infinite membership of items
located at a centroid. The new algorithm assums 1 to be the sum of all memerships of an item. The individual
membership to a specific cluster is the respective percentage towards it.
Also, the algorithm determining the new centroids has been updated to regard the original rank of an item. By
this I aim to reduce the possibility that large topics outweight small topics even though both are very
relevant to the query and appear within the first few results.