Improved Clustering Interface for Project
I changed the interfaces of the cluster interface again and reimplemented the simplefuzzykmeans algorithm accordingly.
The main issue is that instead of returining simply new clusters when planning to add a result, I decided it would be
more appropriate to split this in two - first a new public function was created that enables the creation of new
ClusteredResults. Then an additional method determines what clusters from a provided collection are applicable for
adding this result. The algorithm implementation has been changed to use these new methods.