Project Implementation Backend Refactoring
Refactoring of the project backend. The topic generator was extracted as a separate package thus enabling a strategy-pattern approach. Furthermore the functionality of guessing the cluster based result count was added. The backend should be stable so far. However, the clustering algorithm still has to be reworked some time in the future.